Rename and move files
The generate report module is no longer team specific, so it's been moved for
other teams to more easily use.
